Fix parseFloat in items table price

+++ src/components/MainContent/FeaturedProducts/FeaturedProducts.jsx
@@ -75,7 +75,7 @@ const FeaturesProducts = () => {
                             {getCurrencyData}{" "}
                             {(
                               Math.round(
-                                parseInt(prod.price) *
+                                parseFloat(prod.price) *
                                   parseFloat(actualCurrency()) *
                                   100
                               ) / 100
